+++ src/main/java/org/infernus/idea/checkstyle/util/ConfigurationListeners.java
@@ -1,0 +1,10 @@
+package org.infernus.idea.checkstyle.util;
+
+/**
+ * This enum represents the set of all listener types that can be added to the
+ * Configuration Editor window.
+ */
+public enum ConfigurationListeners {
+  IMPORT_BUTTON_LISTENER, PREVIEW_BUTTON_LISTENER, GENERATE_BUTTON_LISTENER, CATEGORY_SELECT_LISTENER,
+  VISIBLE_RULES_SELECT_LISTENER, ACTIVE_RULES_SELECT_LISTENER;
+}
